+++ src/main/java/pl/coderslab/repository/EmployeeRepository.java
@@ -3,8 +3,6 @@ package pl.coderslab.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import pl.coderslab.model.Employee;
 
-import java.util.List;
-
-public interface EmployeeRepository extends JpaRepository<Employee, Long> {
+public interface EmployeeRepository extends JpaRepository<Employee, Integer> {
 
 }
